Review and refactor import rules
