+++ tests/linter/rules/imports/wrong_import_order/expected_output.txt
@@ -1,0 +1,3 @@
+test.robot:6:12 [W] IMP01 BuiltIn library import 'DateTime' should be placed before 'RequestsLibrary'
+test.robot:7:12 [W] IMP01 BuiltIn library import 'OperatingSystem' should be placed before 'RequestsLibrary'
+test.robot:12:12 [W] IMP01 BuiltIn library import 'BuiltIn' should be placed before 'RequestsLibrary'
