--- tests/linter/rules/misc/builtin_imports_not_sorted/expected_output.txt
@@ -1,4 +1,0 @@
-only_builtins.robot:4:12 [W] 0926 BuiltIn library import 'String' should be placed before 'XML'
-only_builtins.robot:8:12 [W] 0926 BuiltIn library import 'BuiltIn' should be placed before 'String'
-test.robot:3:12 [W] 0926 BuiltIn library import 'Collections' should be placed before 'DateTime'
-test.robot:6:12 [W] 0926 BuiltIn library import 'String' should be placed before 'XML'
